Add explicit conversions to SimScalar literal operators

+++ src/rwe/sim/SimScalar.h
@@ -9,12 +9,12 @@ namespace rwe
 
     constexpr SimScalar operator"" _ss(unsigned long long val)
     {
-        return SimScalar(val);
+        return SimScalar(static_cast<float>(val));
     }
 
     constexpr SimScalar operator"" _ssf(long double val)
     {
-        return SimScalar(val);
+        return SimScalar(static_cast<float>(val));
     }
 
     inline float simScalarToFloat(SimScalar s)
